refactor(core): split impl ThingsDatabase by domain into queries/ and mutations/
The single 4,482-line core.rs impl block is split into 9 domain modules:

  queries/tasks.rs    — get_all_tasks, search_tasks, query_tasks (gated),
                        search_logbook, get_inbox, get_today, get_task_by_uuid
  queries/projects.rs — get_all_projects, get_projects, get_project_by_uuid
  queries/areas.rs    — get_all_areas, get_areas
  queries/tags.rs     — find_tag_by_*, search_tags, get_*_tags,
                        get_tag_completions/statistics, find_duplicate_tags
  mutations/tasks.rs  — create/update/complete/uncomplete/delete_task
  mutations/projects.rs — create/update/complete/delete_project
  mutations/areas.rs  — create/update/delete_area
  mutations/tags.rs   — create/update/delete/merge_tags, add/remove/set_task_tags
  mutations/bulk.rs   — bulk_move/update_dates/complete/delete

core.rs retains only the struct definition plus constructor, pool, and
health-check methods (~1,429 lines, down from 4,482). Zero behaviour change;
all tests pass (622 with advanced-queries, 545 without).
