+++ libs/things3-core/src/database/core.rs
@@ -1,29 +1,17 @@
-#[cfg(any(feature = "advanced-queries", feature = "batch-operations"))]
-use crate::models::TaskFilters;
 use crate::{
     database::{
-        conversions::{naive_date_to_things_timestamp, safe_timestamp_convert},
-        mappers::{map_project_row, map_task_row},
         pool::{
             ComprehensiveHealthStatus, DatabasePoolConfig, PoolHealthStatus, PoolMetrics,
             SqliteOptimizations,
         },
-        query_builders::TaskUpdateBuilder,
         stats::DatabaseStats,
-        validators,
     },
     error::{Result as ThingsResult, ThingsError},
-    models::{
-        Area, CreateTaskRequest, DeleteChildHandling, Project, Task, TaskStatus, TaskType,
-        ThingsId, UpdateTaskRequest,
-    },
 };
-use chrono::{DateTime, NaiveDate, Utc};
-use sqlx::{pool::PoolOptions, Row, SqlitePool};
+use chrono::Utc;
+use sqlx::{pool::PoolOptions, SqlitePool};
 use std::path::Path;
 use tracing::{debug, error, info, instrument};
-#[cfg(any(feature = "advanced-queries", feature = "batch-operations"))]
-use uuid::Uuid;
 
 /// SQLx-based database implementation for Things 3 data
 /// This provides async, Send + Sync compatible database access
@@ -388,3073 +376,31 @@ impl ThingsDatabase {
         })
     }
 
-    /// Get all tasks from the database
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use things3_core::{ThingsDatabase, ThingsError};
-    /// use std::path::Path;
-    ///
-    /// # async fn example() -> Result<(), ThingsError> {
-    /// let db = ThingsDatabase::new(Path::new("/path/to/things.db")).await?;
-    ///
-    /// // Get all tasks
-    /// let tasks = db.get_all_tasks().await?;
-    /// println!("Found {} total tasks", tasks.len());
-    ///
-    /// // Filter tasks by status
-    /// let incomplete: Vec<_> = tasks.iter()
-    ///     .filter(|t| t.status == things3_core::TaskStatus::Incomplete)
-    ///     .collect();
-    /// println!("Found {} incomplete tasks", incomplete.len());
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    #[instrument]
-    pub async fn get_all_tasks(&self) -> ThingsResult<Vec<Task>> {
-        let rows = sqlx::query(
-            r"
-            SELECT 
-                uuid, title, status, type, 
-                start_date, due_date, 
-                project_uuid, area_uuid, 
-                notes, tags, 
-                created, modified
-            FROM TMTask
-            ORDER BY created DESC
-            ",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch tasks: {e}")))?;
-
-        let mut tasks = Vec::new();
-        for row in rows {
-            let task = Task {
-                uuid: ThingsId::from_trusted(row.get::<String, _>("uuid")),
-                title: row.get("title"),
-                status: TaskStatus::from_i32(row.get("status")).unwrap_or(TaskStatus::Incomplete),
-                task_type: TaskType::from_i32(row.get("type")).unwrap_or(TaskType::Todo),
-                start_date: row
-                    .get::<Option<String>, _>("start_date")
-                    .and_then(|s| NaiveDate::parse_from_str(&s, "%Y-%m-%d").ok()),
-                deadline: row
-                    .get::<Option<String>, _>("due_date")
-                    .and_then(|s| NaiveDate::parse_from_str(&s, "%Y-%m-%d").ok()),
-                project_uuid: row
-                    .get::<Option<String>, _>("project_uuid")
-                    .map(ThingsId::from_trusted),
-                area_uuid: row
-                    .get::<Option<String>, _>("area_uuid")
-                    .map(ThingsId::from_trusted),
-                parent_uuid: None, // Not available in this query
-                notes: row.get("notes"),
-                tags: row
-                    .get::<Option<String>, _>("tags")
-                    .map(|s| s.split(',').map(|s| s.trim().to_string()).collect())
-                    .unwrap_or_default(),
-                children: Vec::new(), // Not available in this query
-                created: DateTime::parse_from_rfc3339(&row.get::<String, _>("created"))
-                    .ok()
-                    .map_or_else(Utc::now, |dt| dt.with_timezone(&Utc)),
-                modified: DateTime::parse_from_rfc3339(&row.get::<String, _>("modified"))
-                    .ok()
-                    .map_or_else(Utc::now, |dt| dt.with_timezone(&Utc)),
-                stop_date: None, // Not available in this query context
-            };
-            tasks.push(task);
-        }
-
-        debug!("Fetched {} tasks", tasks.len());
-        Ok(tasks)
-    }
-
-    /// Get all projects (from `TMTask` table where type = 1)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if project data is invalid
-    #[instrument]
-    pub async fn get_all_projects(&self) -> ThingsResult<Vec<Project>> {
-        let rows = sqlx::query(
-            r"
-            SELECT 
-                uuid, title, status, 
-                area, notes, 
-                creationDate, userModificationDate,
-                startDate, deadline
-            FROM TMTask
-            WHERE type = 1 AND trashed = 0
-            ORDER BY creationDate DESC
-            ",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch projects: {e}")))?;
-
-        let mut projects = Vec::new();
-        for row in rows {
-            let project = Project {
-                uuid: ThingsId::from_trusted(row.get::<String, _>("uuid")),
-                title: row.get("title"),
-                status: TaskStatus::from_i32(row.get("status")).unwrap_or(TaskStatus::Incomplete),
-                area_uuid: row
-                    .get::<Option<String>, _>("area")
-                    .map(ThingsId::from_trusted),
-                notes: row.get("notes"),
-                deadline: row
-                    .get::<Option<i64>, _>("deadline")
-                    .and_then(|ts| DateTime::from_timestamp(ts, 0))
-                    .map(|dt| dt.date_naive()),
-                start_date: row
-                    .get::<Option<i64>, _>("startDate")
-                    .and_then(|ts| DateTime::from_timestamp(ts, 0))
-                    .map(|dt| dt.date_naive()),
-                tags: Vec::new(),  // TODO: Load tags separately
-                tasks: Vec::new(), // TODO: Load child tasks separately
-                created: {
-                    let ts_f64 = row.get::<f64, _>("creationDate");
-                    let ts = safe_timestamp_convert(ts_f64);
-                    DateTime::from_timestamp(ts, 0).unwrap_or_else(Utc::now)
-                },
-                modified: {
-                    let ts_f64 = row.get::<f64, _>("userModificationDate");
-                    let ts = safe_timestamp_convert(ts_f64);
-                    DateTime::from_timestamp(ts, 0).unwrap_or_else(Utc::now)
-                },
-            };
-            projects.push(project);
-        }
-
-        debug!("Fetched {} projects", projects.len());
-        Ok(projects)
-    }
-
-    /// Get all areas
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if area data is invalid
-    #[instrument]
-    pub async fn get_all_areas(&self) -> ThingsResult<Vec<Area>> {
-        // Get all areas, not just visible ones (MCP clients may want to see all)
-        let rows = sqlx::query(
-            r"
-            SELECT 
-                uuid, title, visible, `index`
-             FROM TMArea 
-            ORDER BY `index` ASC
-            ",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch areas: {e}")))?;
-
-        let mut areas = Vec::new();
-        for row in rows {
-            let uuid_str: String = row.get("uuid");
-            let area = Area {
-                uuid: ThingsId::from_trusted(uuid_str),
-                title: row.get("title"),
-                notes: None,          // Notes not stored in TMArea table
-                projects: Vec::new(), // TODO: Load projects separately
-                tags: Vec::new(),     // TODO: Load tags separately
-                created: Utc::now(),  // Creation date not available in TMArea
-                modified: Utc::now(), // Modification date not available in TMArea
-            };
-            areas.push(area);
-        }
-
-        debug!("Fetched {} areas", areas.len());
-        Ok(areas)
-    }
-
-    /// Get tasks by status
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    #[instrument]
-    pub async fn get_tasks_by_status(&self, status: TaskStatus) -> ThingsResult<Vec<Task>> {
-        let status_value = status as i32;
-        let rows = sqlx::query(
-            r"
-            SELECT 
-                uuid, title, status, type, 
-                start_date, due_date, 
-                project_uuid, area_uuid, 
-                notes, tags, 
-                created, modified
-             FROM TMTask 
-            WHERE status = ?
-            ORDER BY created DESC
-            ",
-        )
-        .bind(status_value)
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch tasks by status: {e}")))?;
-
-        let mut tasks = Vec::new();
-        for row in rows {
-            let task = Task {
-                uuid: ThingsId::from_trusted(row.get::<String, _>("uuid")),
-                title: row.get("title"),
-                status: TaskStatus::from_i32(row.get("status")).unwrap_or(TaskStatus::Incomplete),
-                task_type: TaskType::from_i32(row.get("type")).unwrap_or(TaskType::Todo),
-                start_date: row
-                    .get::<Option<String>, _>("start_date")
-                    .and_then(|s| NaiveDate::parse_from_str(&s, "%Y-%m-%d").ok()),
-                deadline: row
-                    .get::<Option<String>, _>("due_date")
-                    .and_then(|s| NaiveDate::parse_from_str(&s, "%Y-%m-%d").ok()),
-                project_uuid: row
-                    .get::<Option<String>, _>("project_uuid")
-                    .map(ThingsId::from_trusted),
-                area_uuid: row
-                    .get::<Option<String>, _>("area_uuid")
-                    .map(ThingsId::from_trusted),
-                parent_uuid: None, // Not available in this query
-                notes: row.get("notes"),
-                tags: row
-                    .get::<Option<String>, _>("tags")
-                    .map(|s| s.split(',').map(|s| s.trim().to_string()).collect())
-                    .unwrap_or_default(),
-                children: Vec::new(), // Not available in this query
-                created: DateTime::parse_from_rfc3339(&row.get::<String, _>("created"))
-                    .ok()
-                    .map_or_else(Utc::now, |dt| dt.with_timezone(&Utc)),
-                modified: DateTime::parse_from_rfc3339(&row.get::<String, _>("modified"))
-                    .ok()
-                    .map_or_else(Utc::now, |dt| dt.with_timezone(&Utc)),
-                stop_date: None, // Not available in this query context
-            };
-            tasks.push(task);
-        }
-
-        debug!("Fetched {} tasks with status {:?}", tasks.len(), status);
-        Ok(tasks)
-    }
-
-    /// Search tasks by title or notes
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    #[instrument]
-    pub async fn search_tasks(&self, query: &str) -> ThingsResult<Vec<Task>> {
-        let search_pattern = format!("%{query}%");
-        let rows = sqlx::query(
-            r"
-            SELECT
-                uuid, title, status, type,
-                startDate, deadline, stopDate,
-                project, area, heading,
-                notes,
-                (SELECT GROUP_CONCAT(tg.title, char(31))
-                   FROM TMTaskTag tt
-                   JOIN TMTag tg ON tg.uuid = tt.tags
-                  WHERE tt.tasks = TMTask.uuid) AS tags_csv,
-                creationDate, userModificationDate
-            FROM TMTask
-            WHERE (title LIKE ? OR notes LIKE ?) AND type IN (0, 2) AND trashed = 0
-            ORDER BY creationDate DESC
-            ",
-        )
-        .bind(&search_pattern)
-        .bind(&search_pattern)
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to search tasks: {e}")))?;
-
-        let tasks = rows
-            .iter()
-            .map(map_task_row)
-            .collect::<ThingsResult<Vec<Task>>>()?;
-
-        debug!("Found {} tasks matching query: {}", tasks.len(), query);
-        Ok(tasks)
-    }
-
-    /// Query tasks using a [`TaskFilters`] struct produced by [`crate::query::TaskQueryBuilder`].
-    ///
-    /// All filter fields are optional and combined with AND semantics in SQL.
-    /// Tag and search-query filters are applied in Rust after the SQL query returns
-    /// (Things 3 stores tags as a BLOB). When those post-filters are active,
-    /// `LIMIT`/`OFFSET` is also applied in Rust so pagination counts only
-    /// matching rows; without post-filters it is applied in SQL for efficiency.
-    ///
-    /// Tag matching via `filters.tags` is case-sensitive.
-    ///
-    /// Filtering by [`TaskStatus::Trashed`] queries rows where `trashed = 1`
-    /// rather than adding a `status` condition, matching Things 3's soft-delete
-    /// semantics (trashed rows keep their original status value).
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or task data cannot be mapped.
-    #[cfg(any(feature = "advanced-queries", feature = "batch-operations"))]
-    pub async fn query_tasks(&self, filters: &TaskFilters) -> ThingsResult<Vec<Task>> {
-        self.query_tasks_inner(filters, None).await
-    }
-
-    /// Internal query path that optionally applies a cursor WHERE clause.
-    ///
-    /// `after` is `(seconds_since_unix_epoch, uuid)` of the last-returned
-    /// task. When `Some`, an additional `WHERE` clause restricts results to
-    /// rows strictly older than that anchor in the canonical
-    /// `CAST(creationDate AS INTEGER) DESC, uuid DESC` ordering.
-    ///
-    /// Gated on either `advanced-queries` or `batch-operations` because both
-    /// public surfaces (`query_tasks` and `execute_paged`) share this engine.
-    #[cfg(any(feature = "advanced-queries", feature = "batch-operations"))]
-    pub(crate) async fn query_tasks_inner(
-        &self,
-        filters: &TaskFilters,
-        after: Option<(i64, Uuid)>,
-    ) -> ThingsResult<Vec<Task>> {
-        const COLS: &str = "uuid, title, type, status, notes, startDate, deadline, stopDate, \
-                            creationDate, userModificationDate, project, area, heading, \
-                            (SELECT GROUP_CONCAT(tg.title, char(31)) \
-                               FROM TMTaskTag tt \
-                               JOIN TMTag tg ON tg.uuid = tt.tags \
-                              WHERE tt.tasks = TMTask.uuid) AS tags_csv";
-
-        // Things 3 soft-deletes by setting trashed = 1; the status column is unchanged.
-        // Requesting Trashed means "show trashed rows", not a status = 3 predicate.
-        let trashed_val = i32::from(matches!(filters.status, Some(TaskStatus::Trashed)));
-        let mut conditions: Vec<String> = vec![format!("trashed = {trashed_val}")];
-
-        if let Some(status) = filters.status {
-            let n = match status {
-                TaskStatus::Incomplete => Some(0),
-                TaskStatus::Canceled => Some(2),
-                TaskStatus::Completed => Some(3),
-                TaskStatus::Trashed => None, // handled via trashed = 1 above
-            };
-            if let Some(n) = n {
-                conditions.push(format!("status = {n}"));
-            }
-        }
-
-        if let Some(task_type) = filters.task_type {
-            let n = match task_type {
-                TaskType::Todo => 0,
-                TaskType::Project => 1,
-                TaskType::Heading => 2,
-                TaskType::Area => 3,
-            };
-            conditions.push(format!("type = {n}"));
-        }
-
-        if let Some(ref uuid) = filters.project_uuid {
-            conditions.push(format!("project = '{uuid}'"));
-        }
-
-        if let Some(ref uuid) = filters.area_uuid {
-            conditions.push(format!("area = '{uuid}'"));
-        }
-
-        if let Some(from) = filters.start_date_from {
-            conditions.push(format!(
-                "startDate >= {}",
-                naive_date_to_things_timestamp(from)
-            ));
-        }
-        if let Some(to) = filters.start_date_to {
-            conditions.push(format!(
-                "startDate <= {}",
-                naive_date_to_things_timestamp(to)
-            ));
-        }
-
-        if let Some(from) = filters.deadline_from {
-            conditions.push(format!(
-                "deadline >= {}",
-                naive_date_to_things_timestamp(from)
-            ));
-        }
-        if let Some(to) = filters.deadline_to {
-            conditions.push(format!(
-                "deadline <= {}",
-                naive_date_to_things_timestamp(to)
-            ));
-        }
-
-        if let Some((after_seconds, _)) = after {
-            // Strictly less than the cursor in (truncated_seconds DESC, uuid DESC)
-            // ordering — i.e. older second, or same second with smaller uuid.
-            // Casting to INTEGER matches the precision of `Task::created`,
-            // which is reconstructed at second precision when reading rows.
-            // UUID is bound as a parameter (?) rather than interpolated for
-            // consistency with the rest of the codebase's query practices.
-            conditions.push(format!(
-                "(CAST(creationDate AS INTEGER) < {after_seconds} \
-                 OR (CAST(creationDate AS INTEGER) = {after_seconds} AND uuid < ?))"
-            ));
-        }
-
-        let where_clause = conditions.join(" AND ");
-        // ORDER BY uses the truncated-second value so it agrees with the
-        // cursor pagination logic (which compares at second precision because
-        // `Task::created` is reconstructed at second precision). `uuid DESC` is
-        // a deterministic tiebreak within the same second.
-        let mut sql = format!(
-            "SELECT {COLS} FROM TMTask WHERE {where_clause} \
-             ORDER BY CAST(creationDate AS INTEGER) DESC, uuid DESC"
-        );
-
-        // When tags or search_query are active, LIMIT/OFFSET must be applied in Rust
-        // after post-filtering, because SQL LIMIT would count non-matching rows.
-        let has_post_filters =
-            filters.tags.as_ref().is_some_and(|t| !t.is_empty()) || filters.search_query.is_some();
-
-        if !has_post_filters {
-            match (filters.limit, filters.offset) {
-                (Some(limit), Some(offset)) => {
-                    sql.push_str(&format!(" LIMIT {limit} OFFSET {offset}"));
-                }
-                (Some(limit), None) => {
-                    sql.push_str(&format!(" LIMIT {limit}"));
-                }
-                (None, Some(offset)) => {
-                    // SQLite requires LIMIT when OFFSET is used; -1 means unlimited
-                    sql.push_str(&format!(" LIMIT -1 OFFSET {offset}"));
-                }
-                (None, None) => {}
-            }
-        }
-
-        let rows = if let Some((_, after_uuid)) = after {
-            sqlx::query(&sql)
-                .bind(after_uuid.to_string())
-                .fetch_all(&self.pool)
-                .await
-        } else {
-            sqlx::query(&sql).fetch_all(&self.pool).await
-        }
-        .map_err(|e| ThingsError::unknown(format!("Failed to query tasks: {e}")))?;
-
-        let mut tasks = rows
-            .iter()
-            .map(map_task_row)
-            .collect::<ThingsResult<Vec<Task>>>()?;
-
-        if let Some(ref filter_tags) = filters.tags {
-            if !filter_tags.is_empty() {
-                tasks.retain(|task| filter_tags.iter().all(|f| task.tags.contains(f)));
-            }
-        }
-
-        if let Some(ref q) = filters.search_query {
-            let q_lower = q.to_lowercase();
-            tasks.retain(|task| {
-                task.title.to_lowercase().contains(&q_lower)
-                    || task
-                        .notes
-                        .as_deref()
-                        .unwrap_or("")
-                        .to_lowercase()
-                        .contains(&q_lower)
-            });
-        }
-
-        if has_post_filters {
-            let offset = filters.offset.unwrap_or(0);
-            tasks = tasks.into_iter().skip(offset).collect();
-            if let Some(limit) = filters.limit {
-                tasks.truncate(limit);
-            }
-        }
-
-        Ok(tasks)
-    }
-
-    /// Search completed tasks in the logbook
-    ///
-    /// Returns completed tasks matching the provided filters.
-    /// All filters are optional and can be combined.
-    ///
-    /// # Parameters
-    ///
-    /// - `search_text`: Search in task titles and notes (case-insensitive)
-    /// - `from_date`: Start date for completion date range
-    /// - `to_date`: End date for completion date range
-    /// - `project_uuid`: Filter by project UUID
-    /// - `area_uuid`: Filter by area UUID
-    /// - `tags`: Filter by tags (all tags must match)
-    /// - `limit`: Maximum number of results (default: 50)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    #[allow(clippy::too_many_arguments)]
-    #[instrument(skip(self))]
-    pub async fn search_logbook(
-        &self,
-        search_text: Option<String>,
-        from_date: Option<NaiveDate>,
-        to_date: Option<NaiveDate>,
-        project_uuid: Option<ThingsId>,
-        area_uuid: Option<ThingsId>,
-        tags: Option<Vec<String>>,
-        limit: Option<u32>,
-        offset: Option<u32>,
-    ) -> ThingsResult<Vec<Task>> {
-        // Apply limit and offset
-        let result_limit = limit.unwrap_or(50).min(500);
-        let result_offset = offset.unwrap_or(0);
-
-        // Build and execute query based on filters
-        // type = 0 (Todo) is intentional here: headings (type=2) have no stopDate and
-        // cannot appear in a stop-date-ordered logbook.
-        let rows = if let Some(ref text) = search_text {
-            let pattern = format!("%{text}%");
-            let mut q = String::from(
-                "SELECT uuid, title, status, type, startDate, deadline, stopDate, project, area, heading, notes, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv, creationDate, userModificationDate FROM TMTask WHERE status = 3 AND trashed = 0 AND type = 0",
-            );
-            q.push_str(" AND (title LIKE ? OR notes LIKE ?)");
-
-            if let Some(date) = from_date {
-                // stopDate is stored as Unix timestamp (seconds since 1970-01-01)
-                let date_time = date.and_hms_opt(0, 0, 0).unwrap().and_utc();
-                let timestamp = date_time.timestamp() as f64;
-                q.push_str(&format!(" AND stopDate >= {}", timestamp));
-            }
-
-            if let Some(date) = to_date {
-                // Include tasks completed on to_date by adding 1 day
-                let end_date = date + chrono::Duration::days(1);
-                let date_time = end_date.and_hms_opt(0, 0, 0).unwrap().and_utc();
-                let timestamp = date_time.timestamp() as f64;
-                q.push_str(&format!(" AND stopDate < {}", timestamp));
-            }
-
-            if let Some(ref id) = project_uuid {
-                q.push_str(&format!(" AND project = '{}'", id));
-            }
-
-            if let Some(ref id) = area_uuid {
-                q.push_str(&format!(" AND area = '{}'", id));
-            }
-
-            q.push_str(&format!(
-                " ORDER BY stopDate DESC LIMIT {result_limit} OFFSET {result_offset}"
-            ));
-
-            sqlx::query(&q)
-                .bind(&pattern)
-                .bind(&pattern)
-                .fetch_all(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to search logbook: {e}")))?
-        } else {
-            let mut q = String::from(
-                "SELECT uuid, title, status, type, startDate, deadline, stopDate, project, area, heading, notes, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv, creationDate, userModificationDate FROM TMTask WHERE status = 3 AND trashed = 0 AND type = 0",
-            );
-
-            if let Some(date) = from_date {
-                // stopDate is stored as Unix timestamp (seconds since 1970-01-01)
-                let date_time = date.and_hms_opt(0, 0, 0).unwrap().and_utc();
-                let timestamp = date_time.timestamp() as f64;
-                q.push_str(&format!(" AND stopDate >= {}", timestamp));
-            }
-
-            if let Some(date) = to_date {
-                // Include tasks completed on to_date by adding 1 day
-                let end_date = date + chrono::Duration::days(1);
-                let date_time = end_date.and_hms_opt(0, 0, 0).unwrap().and_utc();
-                let timestamp = date_time.timestamp() as f64;
-                q.push_str(&format!(" AND stopDate < {}", timestamp));
-            }
-
-            if let Some(ref id) = project_uuid {
-                q.push_str(&format!(" AND project = '{}'", id));
-            }
-
-            if let Some(ref id) = area_uuid {
-                q.push_str(&format!(" AND area = '{}'", id));
-            }
-
-            q.push_str(&format!(
-                " ORDER BY stopDate DESC LIMIT {result_limit} OFFSET {result_offset}"
-            ));
-
-            sqlx::query(&q)
-                .fetch_all(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to search logbook: {e}")))?
-        };
-
-        // Filter by tags if provided
-        let mut tasks = rows
-            .iter()
-            .map(map_task_row)
-            .collect::<ThingsResult<Vec<Task>>>()?;
-
-        if let Some(ref filter_tags) = tags {
-            if !filter_tags.is_empty() {
-                tasks.retain(|task| {
-                    // Check if task has all required tags
-                    filter_tags
-                        .iter()
-                        .all(|filter_tag| task.tags.contains(filter_tag))
-                });
-            }
-        }
-
-        debug!("Found {} completed tasks in logbook", tasks.len());
-        Ok(tasks)
-    }
-
-    /// Get inbox tasks (incomplete tasks without project)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    #[instrument(skip(self))]
-    pub async fn get_inbox(&self, limit: Option<usize>) -> ThingsResult<Vec<Task>> {
-        let query = if let Some(limit) = limit {
-            format!("SELECT uuid, title, type, status, notes, startDate, deadline, stopDate, creationDate, userModificationDate, project, area, heading, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv FROM TMTask WHERE type IN (0, 2) AND status = 0 AND project IS NULL AND trashed = 0 ORDER BY creationDate DESC LIMIT {limit}")
-        } else {
-            "SELECT uuid, title, type, status, notes, startDate, deadline, stopDate, creationDate, userModificationDate, project, area, heading, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv FROM TMTask WHERE type IN (0, 2) AND status = 0 AND project IS NULL AND trashed = 0 ORDER BY creationDate DESC"
-                .to_string()
-        };
-
-        let rows = sqlx::query(&query)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to fetch inbox tasks: {e}")))?;
-
-        let tasks = rows
-            .iter()
-            .map(map_task_row)
-            .collect::<ThingsResult<Vec<Task>>>()?;
-
-        Ok(tasks)
-    }
-
-    /// Get today's tasks (incomplete tasks due today or started today)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if task data is invalid
-    ///
-    /// # Panics
-    ///
-    /// Panics if the current date cannot be converted to a valid time with hours, minutes, and seconds
-    #[instrument(skip(self))]
-    pub async fn get_today(&self, limit: Option<usize>) -> ThingsResult<Vec<Task>> {
-        // Things 3 uses the `todayIndex` column to mark tasks that appear in "Today"
-        // A task is in "Today" if todayIndex IS NOT NULL AND todayIndex != 0
-        let query = if let Some(limit) = limit {
-            format!(
-                "SELECT uuid, title, type, status, notes, startDate, deadline, stopDate, creationDate, userModificationDate, project, area, heading, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv FROM TMTask WHERE status = 0 AND todayIndex IS NOT NULL AND todayIndex != 0 AND trashed = 0 ORDER BY todayIndex ASC LIMIT {limit}"
-            )
-        } else {
-            "SELECT uuid, title, type, status, notes, startDate, deadline, stopDate, creationDate, userModificationDate, project, area, heading, (SELECT GROUP_CONCAT(tg.title, char(31)) FROM TMTaskTag tt JOIN TMTag tg ON tg.uuid = tt.tags WHERE tt.tasks = TMTask.uuid) AS tags_csv FROM TMTask WHERE status = 0 AND todayIndex IS NOT NULL AND todayIndex != 0 AND trashed = 0 ORDER BY todayIndex ASC".to_string()
-        };
-
-        let rows = sqlx::query(&query)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to fetch today's tasks: {e}")))?;
-
-        let tasks = rows
-            .iter()
-            .map(map_task_row)
-            .collect::<ThingsResult<Vec<Task>>>()?;
-
-        Ok(tasks)
-    }
-
-    /// Get all projects (alias for `get_all_projects` for compatibility)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if project data is invalid
-    #[instrument(skip(self))]
-    pub async fn get_projects(&self, limit: Option<usize>) -> ThingsResult<Vec<Project>> {
-        let _ = limit; // Currently unused but kept for API compatibility
-        self.get_all_projects().await
-    }
-
-    /// Get all areas (alias for `get_all_areas` for compatibility)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails or if area data is invalid
-    #[instrument(skip(self))]
-    pub async fn get_areas(&self) -> ThingsResult<Vec<Area>> {
-        self.get_all_areas().await
-    }
-
-    /// Create a new task in the database
-    ///
-    /// Validates:
-    /// - Project UUID exists if provided
-    /// - Area UUID exists if provided
-    /// - Parent task UUID exists if provided
-    /// - Date range (deadline >= start_date)
-    ///
-    /// Returns the UUID of the created task
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use things3_core::{ThingsDatabase, CreateTaskRequest, ThingsError};
-    /// use std::path::Path;
-    /// use chrono::NaiveDate;
-    ///
-    /// # async fn example() -> Result<(), ThingsError> {
-    /// let db = ThingsDatabase::new(Path::new("/path/to/things.db")).await?;
-    ///
-    /// // Create a simple task
-    /// let request = CreateTaskRequest {
-    ///     title: "Buy groceries".to_string(),
-    ///     notes: Some("Milk, eggs, bread".to_string()),
-    ///     deadline: Some(NaiveDate::from_ymd_opt(2024, 12, 31).unwrap()),
-    ///     start_date: None,
-    ///     project_uuid: None,
-    ///     area_uuid: None,
-    ///     parent_uuid: None,
-    ///     tags: None,
-    ///     task_type: None,
-    ///     status: None,
-    /// };
-    ///
-    /// let task_uuid = db.create_task(request).await?;
-    /// println!("Created task with UUID: {}", task_uuid);
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if validation fails or if the database insert fails
-    #[instrument(skip(self))]
-    pub async fn create_task(&self, request: CreateTaskRequest) -> ThingsResult<ThingsId> {
-        // Validate date range (deadline must be >= start_date)
-        crate::database::validate_date_range(request.start_date, request.deadline)?;
-
-        // Generate ID for new task
-        let id = ThingsId::new_things_native();
-
-        // Validate referenced entities
-        if let Some(project_uuid) = &request.project_uuid {
-            validators::validate_project_exists(&self.pool, project_uuid).await?;
-        }
-
-        if let Some(area_uuid) = &request.area_uuid {
-            validators::validate_area_exists(&self.pool, area_uuid).await?;
-        }
-
-        if let Some(parent_uuid) = &request.parent_uuid {
-            validators::validate_task_exists(&self.pool, parent_uuid).await?;
-        }
-
-        // Convert dates to Things 3 format (seconds since 2001-01-01)
-        let start_date_ts = request.start_date.map(naive_date_to_things_timestamp);
-        let deadline_ts = request.deadline.map(naive_date_to_things_timestamp);
-
-        // Get current timestamp for creation/modification dates
-        let now = Utc::now().timestamp() as f64;
-
-        // Insert into TMTask table
-        sqlx::query(
-            r"
-            INSERT INTO TMTask (
-                uuid, title, type, status, notes,
-                startDate, deadline, project, area, heading,
-                creationDate, userModificationDate,
-                trashed
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            ",
-        )
-        .bind(id.as_str())
-        .bind(&request.title)
-        .bind(request.task_type.unwrap_or(TaskType::Todo) as i32)
-        .bind(request.status.unwrap_or(TaskStatus::Incomplete) as i32)
-        .bind(request.notes.as_ref())
-        .bind(start_date_ts)
-        .bind(deadline_ts)
-        .bind(request.project_uuid.map(|u| u.into_string()))
-        .bind(request.area_uuid.map(|u| u.into_string()))
-        .bind(request.parent_uuid.map(|u| u.into_string()))
-        .bind(now)
-        .bind(now)
-        .bind(0) // not trashed
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to create task: {e}")))?;
-
-        // Handle tags via TMTaskTag
-        if let Some(tags) = request.tags {
-            self.set_task_tags(&id, tags).await?;
-        }
-
-        info!("Created task with UUID: {}", id);
-        Ok(id)
-    }
-
-    /// Create a new project
-    ///
-    /// Projects are tasks with type = 1 in the TMTask table
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if validation fails or the database insert fails
-    #[instrument(skip(self))]
-    pub async fn create_project(
-        &self,
-        request: crate::models::CreateProjectRequest,
-    ) -> ThingsResult<ThingsId> {
-        // Validate date range (deadline must be >= start_date)
-        crate::database::validate_date_range(request.start_date, request.deadline)?;
-
-        // Generate ID for new project
-        let id = ThingsId::new_things_native();
-
-        // Validate area if provided
-        if let Some(area_uuid) = &request.area_uuid {
-            validators::validate_area_exists(&self.pool, area_uuid).await?;
-        }
-
-        // Convert dates to Things 3 format (seconds since 2001-01-01)
-        let start_date_ts = request.start_date.map(naive_date_to_things_timestamp);
-        let deadline_ts = request.deadline.map(naive_date_to_things_timestamp);
-
-        // Get current timestamp for creation/modification dates
-        let now = Utc::now().timestamp() as f64;
-
-        // Insert into TMTask table with type = 1 (project)
-        sqlx::query(
-            r"
-            INSERT INTO TMTask (
-                uuid, title, type, status, notes,
-                startDate, deadline, project, area, heading,
-                creationDate, userModificationDate,
-                trashed
-            ) VALUES (?, ?, 1, 0, ?, ?, ?, NULL, ?, NULL, ?, ?, 0)
-            ",
-        )
-        .bind(id.as_str())
-        .bind(&request.title)
-        .bind(request.notes.as_ref())
-        .bind(start_date_ts)
-        .bind(deadline_ts)
-        .bind(request.area_uuid.map(|u| u.into_string()))
-        .bind(now)
-        .bind(now)
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to create project: {e}")))?;
-
-        // Handle tags via TMTaskTag
-        if let Some(tags) = request.tags {
-            self.set_task_tags(&id, tags).await?;
-        }
-
-        info!("Created project with UUID: {}", id);
-        Ok(id)
-    }
-
-    /// Update an existing task
-    ///
-    /// Only updates fields that are provided (Some(_))
-    /// Validates existence of referenced entities
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task doesn't exist, validation fails, or the database update fails
-    #[instrument(skip(self))]
-    pub async fn update_task(&self, request: UpdateTaskRequest) -> ThingsResult<()> {
-        // Verify task exists
-        validators::validate_task_exists(&self.pool, &request.uuid).await?;
-
-        // Validate dates if either is being updated
-        if request.start_date.is_some() || request.deadline.is_some() {
-            // Get current task to merge dates
-            if let Some(current_task) = self.get_task_by_uuid(&request.uuid).await? {
-                let final_start = request.start_date.or(current_task.start_date);
-                let final_deadline = request.deadline.or(current_task.deadline);
-                crate::database::validate_date_range(final_start, final_deadline)?;
-            }
-        }
-
-        // Validate referenced entities if being updated
-        if let Some(project_uuid) = &request.project_uuid {
-            validators::validate_project_exists(&self.pool, project_uuid).await?;
-        }
-
-        if let Some(area_uuid) = &request.area_uuid {
-            validators::validate_area_exists(&self.pool, area_uuid).await?;
-        }
-
-        // Use the TaskUpdateBuilder to construct the query
-        let builder = TaskUpdateBuilder::from_request(&request);
-
-        // If no fields to update, just return (modification date will still be updated)
-        if builder.is_empty() {
-            return Ok(());
-        }
-
-        let query_string = builder.build_query_string();
-        let mut q = sqlx::query(&query_string);
-
-        // Bind values in the same order as the builder added fields
-        if let Some(title) = &request.title {
-            q = q.bind(title);
-        }
-
-        if let Some(notes) = &request.notes {
-            q = q.bind(notes);
-        }
-
-        if let Some(start_date) = request.start_date {
-            q = q.bind(naive_date_to_things_timestamp(start_date));
-        }
-
-        if let Some(deadline) = request.deadline {
-            q = q.bind(naive_date_to_things_timestamp(deadline));
-        }
-
-        if let Some(status) = request.status {
-            q = q.bind(status as i32);
-        }
-
-        if let Some(project_uuid) = request.project_uuid {
-            q = q.bind(project_uuid.into_string());
-        }
-
-        if let Some(area_uuid) = request.area_uuid {
-            q = q.bind(area_uuid.into_string());
-        }
-
-        // Bind modification date and UUID (always added by builder)
-        let now = Utc::now().timestamp() as f64;
-        q = q.bind(now).bind(request.uuid.as_str());
-
-        q.execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to update task: {e}")))?;
-
-        // Handle tags via TMTaskTag (separate from the UPDATE query)
-        if let Some(tags) = request.tags {
-            self.set_task_tags(&request.uuid, tags).await?;
-        }
-
-        info!("Updated task with UUID: {}", request.uuid);
-        Ok(())
-    }
-
-    /// Get a single project by UUID
-    ///
-    /// Returns `None` if the project doesn't exist or is trashed
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_project_by_uuid(&self, id: &ThingsId) -> ThingsResult<Option<Project>> {
-        let row = sqlx::query(
-            r"
-            SELECT
-                uuid, title, status,
-                area, notes,
-                creationDate, userModificationDate,
-                startDate, deadline,
-                trashed, type
-            FROM TMTask
-            WHERE uuid = ? AND type = 1
-            ",
-        )
-        .bind(id.as_str())
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch project: {e}")))?;
-
-        if let Some(row) = row {
-            let trashed: i64 = row.get("trashed");
-            if trashed == 1 {
-                return Ok(None);
-            }
-            Ok(Some(map_project_row(&row)))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Update an existing project
-    ///
-    /// Only updates fields that are provided (Some(_))
-    /// Validates existence and that the entity is a project (type = 1)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the project doesn't exist, validation fails, or the database update fails
-    #[instrument(skip(self))]
-    pub async fn update_project(
-        &self,
-        request: crate::models::UpdateProjectRequest,
-    ) -> ThingsResult<()> {
-        // Verify project exists (type = 1, trashed = 0)
-        validators::validate_project_exists(&self.pool, &request.uuid).await?;
-
-        // Validate dates if either is being updated
-        if request.start_date.is_some() || request.deadline.is_some() {
-            // Fetch current project to merge dates
-            if let Some(current_project) = self.get_project_by_uuid(&request.uuid).await? {
-                let final_start = request.start_date.or(current_project.start_date);
-                let final_deadline = request.deadline.or(current_project.deadline);
-                crate::database::validate_date_range(final_start, final_deadline)?;
-            }
-        }
-
-        // Validate area if being updated
-        if let Some(area_uuid) = &request.area_uuid {
-            validators::validate_area_exists(&self.pool, area_uuid).await?;
-        }
-
-        // Build dynamic query using TaskUpdateBuilder
-        let mut builder = TaskUpdateBuilder::new();
-
-        // Add fields to update
-        if request.title.is_some() {
-            builder = builder.add_field("title");
-        }
-        if request.notes.is_some() {
-            builder = builder.add_field("notes");
-        }
-        if request.start_date.is_some() {
-            builder = builder.add_field("startDate");
-        }
-        if request.deadline.is_some() {
-            builder = builder.add_field("deadline");
-        }
-        if request.area_uuid.is_some() {
-            builder = builder.add_field("area");
-        }
-
-        // If nothing to update (tags-only changes still need to proceed for TMTaskTag)
-        let has_db_fields = !builder.is_empty();
-
-        if has_db_fields {
-            // Build query string
-            let query_str = builder.build_query_string();
-            let mut q = sqlx::query(&query_str);
-
-            // Bind values in the same order they were added to the builder
-            if let Some(ref title) = request.title {
-                q = q.bind(title);
-            }
-            if let Some(ref notes) = request.notes {
-                q = q.bind(notes);
-            }
-            if let Some(start_date) = request.start_date {
-                q = q.bind(naive_date_to_things_timestamp(start_date));
-            }
-            if let Some(deadline) = request.deadline {
-                q = q.bind(naive_date_to_things_timestamp(deadline));
-            }
-            if let Some(area_uuid) = request.area_uuid {
-                q = q.bind(area_uuid.into_string());
-            }
-
-            // Bind modification date and UUID (always added by builder)
-            let now = Utc::now().timestamp() as f64;
-            q = q.bind(now).bind(request.uuid.as_str());
-
-            q.execute(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to update project: {e}")))?;
-        } else if request.tags.is_none() {
-            // Nothing to update at all
-            return Ok(());
-        }
-
-        // Handle tags via TMTaskTag (separate from the UPDATE query)
-        if let Some(tags) = request.tags {
-            self.set_task_tags(&request.uuid, tags).await?;
-        }
-
-        info!("Updated project with UUID: {}", request.uuid);
-        Ok(())
-    }
-
-    /// Get a task by its UUID
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task does not exist or if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_task_by_uuid(&self, id: &ThingsId) -> ThingsResult<Option<Task>> {
-        let row = sqlx::query(
-            r"
-            SELECT
-                uuid, title, status, type,
-                startDate, deadline, stopDate,
-                project, area, heading,
-                notes, (SELECT GROUP_CONCAT(tg.title, char(31))
-                          FROM TMTaskTag tt
-                          JOIN TMTag tg ON tg.uuid = tt.tags
-                         WHERE tt.tasks = TMTask.uuid) AS tags_csv,
-                creationDate, userModificationDate,
-                trashed
-            FROM TMTask
-            WHERE uuid = ?
-            ",
-        )
-        .bind(id.as_str())
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to fetch task: {e}")))?;
-
-        if let Some(row) = row {
-            // Check if trashed
-            let trashed: i64 = row.get("trashed");
-            if trashed == 1 {
-                return Ok(None); // Return None for trashed tasks
-            }
-
-            // Use the centralized mapper
-            let task = map_task_row(&row)?;
-            Ok(Some(task))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Mark a task as completed
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task does not exist or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn complete_task(&self, id: &ThingsId) -> ThingsResult<()> {
-        // Verify task exists
-        validators::validate_task_exists(&self.pool, id).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        sqlx::query(
-            "UPDATE TMTask SET status = 3, stopDate = ?, userModificationDate = ? WHERE uuid = ?",
-        )
-        .bind(now)
-        .bind(now)
-        .bind(id.as_str())
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to complete task: {e}")))?;
-
-        info!("Completed task with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Mark a completed task as incomplete
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task does not exist or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn uncomplete_task(&self, id: &ThingsId) -> ThingsResult<()> {
-        // Verify task exists
-        validators::validate_task_exists(&self.pool, id).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        sqlx::query(
-            "UPDATE TMTask SET status = 0, stopDate = NULL, userModificationDate = ? WHERE uuid = ?",
-        )
-        .bind(now)
-        .bind(id.as_str())
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to uncomplete task: {e}")))?;
-
-        info!("Uncompleted task with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Complete a project and optionally handle its child tasks
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the project doesn't exist or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn complete_project(
-        &self,
-        id: &ThingsId,
-        child_handling: crate::models::ProjectChildHandling,
-    ) -> ThingsResult<()> {
-        // Verify project exists
-        validators::validate_project_exists(&self.pool, id).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        // Handle child tasks based on the handling mode
-        match child_handling {
-            crate::models::ProjectChildHandling::Error => {
-                // Check if project has children
-                let child_count: i64 = sqlx::query_scalar(
-                    "SELECT COUNT(*) FROM TMTask WHERE project = ? AND trashed = 0",
-                )
-                .bind(id.as_str())
-                .fetch_one(&self.pool)
-                .await
-                .map_err(|e| {
-                    ThingsError::unknown(format!("Failed to check for child tasks: {e}"))
-                })?;
-
-                if child_count > 0 {
-                    return Err(ThingsError::unknown(format!(
-                        "Project {} has {} child task(s). Use cascade or orphan mode to complete.",
-                        id, child_count
-                    )));
-                }
-            }
-            crate::models::ProjectChildHandling::Cascade => {
-                // Complete all child tasks
-                sqlx::query(
-                    "UPDATE TMTask SET status = 3, stopDate = ?, userModificationDate = ? WHERE project = ? AND trashed = 0",
-                )
-                .bind(now)
-                .bind(now)
-                .bind(id.as_str())
-                .execute(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to complete child tasks: {e}")))?;
-            }
-            crate::models::ProjectChildHandling::Orphan => {
-                // Move child tasks to inbox (set project to NULL)
-                sqlx::query(
-                    "UPDATE TMTask SET project = NULL, userModificationDate = ? WHERE project = ? AND trashed = 0",
-                )
-                .bind(now)
-                .bind(id.as_str())
-                .execute(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to orphan child tasks: {e}")))?;
-            }
-        }
-
-        // Complete the project
-        sqlx::query(
-            "UPDATE TMTask SET status = 3, stopDate = ?, userModificationDate = ? WHERE uuid = ?",
-        )
-        .bind(now)
-        .bind(now)
-        .bind(id.as_str())
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to complete project: {e}")))?;
-
-        info!("Completed project with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Soft delete a task (set trashed flag)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task does not exist, if child handling fails, or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn delete_task(
-        &self,
-        id: &ThingsId,
-        child_handling: DeleteChildHandling,
-    ) -> ThingsResult<()> {
-        // Verify task exists
-        validators::validate_task_exists(&self.pool, id).await?;
-
-        // Check for child tasks
-        let children = sqlx::query("SELECT uuid FROM TMTask WHERE heading = ? AND trashed = 0")
-            .bind(id.as_str())
-            .fetch_all(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to query child tasks: {e}")))?;
-
-        let has_children = !children.is_empty();
-
-        if has_children {
-            match child_handling {
-                DeleteChildHandling::Error => {
-                    return Err(ThingsError::unknown(format!(
-                        "Task {} has {} child task(s). Use cascade or orphan mode to delete.",
-                        id,
-                        children.len()
-                    )));
-                }
-                DeleteChildHandling::Cascade => {
-                    // Delete all children
-                    let now = Utc::now().timestamp() as f64;
-                    for child_row in &children {
-                        let child_uuid: String = child_row.get("uuid");
-                        sqlx::query(
-                            "UPDATE TMTask SET trashed = 1, userModificationDate = ? WHERE uuid = ?",
-                        )
-                        .bind(now)
-                        .bind(&child_uuid)
-                        .execute(&self.pool)
-                        .await
-                        .map_err(|e| {
-                            ThingsError::unknown(format!("Failed to delete child task: {e}"))
-                        })?;
-                    }
-                    info!("Cascade deleted {} child task(s)", children.len());
-                }
-                DeleteChildHandling::Orphan => {
-                    // Clear parent reference for children
-                    let now = Utc::now().timestamp() as f64;
-                    for child_row in &children {
-                        let child_uuid: String = child_row.get("uuid");
-                        sqlx::query(
-                            "UPDATE TMTask SET heading = NULL, userModificationDate = ? WHERE uuid = ?",
-                        )
-                        .bind(now)
-                        .bind(&child_uuid)
-                        .execute(&self.pool)
-                        .await
-                        .map_err(|e| {
-                            ThingsError::unknown(format!("Failed to orphan child task: {e}"))
-                        })?;
-                    }
-                    info!("Orphaned {} child task(s)", children.len());
-                }
-            }
-        }
-
-        // Delete the parent task
-        let now = Utc::now().timestamp() as f64;
-        sqlx::query("UPDATE TMTask SET trashed = 1, userModificationDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to delete task: {e}")))?;
-
-        info!("Deleted task with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Soft delete a project and handle its child tasks
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the project doesn't exist, if child handling fails, or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn delete_project(
-        &self,
-        id: &ThingsId,
-        child_handling: crate::models::ProjectChildHandling,
-    ) -> ThingsResult<()> {
-        // Verify project exists
-        validators::validate_project_exists(&self.pool, id).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        // Handle child tasks based on the handling mode
-        match child_handling {
-            crate::models::ProjectChildHandling::Error => {
-                // Check if project has children
-                let child_count: i64 = sqlx::query_scalar(
-                    "SELECT COUNT(*) FROM TMTask WHERE project = ? AND trashed = 0",
-                )
-                .bind(id.as_str())
-                .fetch_one(&self.pool)
-                .await
-                .map_err(|e| {
-                    ThingsError::unknown(format!("Failed to check for child tasks: {e}"))
-                })?;
-
-                if child_count > 0 {
-                    return Err(ThingsError::unknown(format!(
-                        "Project {} has {} child task(s). Use cascade or orphan mode to delete.",
-                        id, child_count
-                    )));
-                }
-            }
-            crate::models::ProjectChildHandling::Cascade => {
-                // Delete all child tasks
-                sqlx::query(
-                    "UPDATE TMTask SET trashed = 1, userModificationDate = ? WHERE project = ? AND trashed = 0",
-                )
-                .bind(now)
-                .bind(id.as_str())
-                .execute(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to delete child tasks: {e}")))?;
-            }
-            crate::models::ProjectChildHandling::Orphan => {
-                // Move child tasks to inbox (set project to NULL)
-                sqlx::query(
-                    "UPDATE TMTask SET project = NULL, userModificationDate = ? WHERE project = ? AND trashed = 0",
-                )
-                .bind(now)
-                .bind(id.as_str())
-                .execute(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to orphan child tasks: {e}")))?;
-            }
-        }
-
-        // Delete the project
-        sqlx::query("UPDATE TMTask SET trashed = 1, userModificationDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to delete project: {e}")))?;
-
-        info!("Deleted project with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Create a new area
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database insert fails
-    #[instrument(skip(self))]
-    pub async fn create_area(
-        &self,
-        request: crate::models::CreateAreaRequest,
-    ) -> ThingsResult<ThingsId> {
-        // Generate ID for new area
-        let id = ThingsId::new_things_native();
-
-        // Get current timestamp for creation/modification dates
-        let now = Utc::now().timestamp() as f64;
-
-        // Calculate next index (max + 1)
-        let max_index: Option<i64> = sqlx::query_scalar("SELECT MAX(`index`) FROM TMArea")
-            .fetch_one(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to get max area index: {e}")))?;
-
-        let next_index = max_index.unwrap_or(-1) + 1;
-
-        // Insert into TMArea table
-        sqlx::query(
-            r"
-            INSERT INTO TMArea (
-                uuid, title, visible, `index`,
-                creationDate, userModificationDate
-            ) VALUES (?, ?, 1, ?, ?, ?)
-            ",
-        )
-        .bind(id.as_str())
-        .bind(&request.title)
-        .bind(next_index)
-        .bind(now)
-        .bind(now)
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to create area: {e}")))?;
-
-        info!("Created area with UUID: {}", id);
-        Ok(id)
-    }
-
-    /// Update an existing area
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the area doesn't exist or if the database update fails
-    #[instrument(skip(self))]
-    pub async fn update_area(&self, request: crate::models::UpdateAreaRequest) -> ThingsResult<()> {
-        // Verify area exists
-        validators::validate_area_exists(&self.pool, &request.uuid).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        sqlx::query("UPDATE TMArea SET title = ?, userModificationDate = ? WHERE uuid = ?")
-            .bind(&request.title)
-            .bind(now)
-            .bind(request.uuid.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to update area: {e}")))?;
-
-        info!("Updated area with UUID: {}", request.uuid);
-        Ok(())
-    }
-
-    /// Delete an area
-    ///
-    /// Hard delete (areas don't have a trashed field)
-    /// Orphans all projects in the area by setting their area to NULL
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the area doesn't exist or if the database delete fails
-    #[instrument(skip(self))]
-    pub async fn delete_area(&self, id: &ThingsId) -> ThingsResult<()> {
-        // Verify area exists
-        validators::validate_area_exists(&self.pool, id).await?;
-
-        let now = Utc::now().timestamp() as f64;
-
-        // Orphan all projects in this area (set area to NULL)
-        sqlx::query(
-            "UPDATE TMTask SET area = NULL, userModificationDate = ? WHERE area = ? AND type = 1 AND trashed = 0",
-        )
-        .bind(now)
-        .bind(id.as_str())
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to orphan projects in area: {e}")))?;
-
-        // Delete the area (hard delete)
-        sqlx::query("DELETE FROM TMArea WHERE uuid = ?")
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to delete area: {e}")))?;
-
-        info!("Deleted area with UUID: {}", id);
-        Ok(())
-    }
-
     // ========================================================================
     // TAG OPERATIONS (with smart duplicate prevention)
     // ========================================================================
-
-    /// Find a tag by normalized title (exact match, case-insensitive)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn find_tag_by_normalized_title(
-        &self,
-        normalized: &str,
-    ) -> ThingsResult<Option<crate::models::Tag>> {
-        let row = sqlx::query(
-            "SELECT uuid, title, shortcut, parent, usedDate
-             FROM TMTag
-             WHERE LOWER(title) = LOWER(?)",
-        )
-        .bind(normalized)
-        .fetch_optional(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to find tag by title: {e}")))?;
-
-        if let Some(row) = row {
-            let uuid_str: String = row.get("uuid");
-            let title: String = row.get("title");
-            let shortcut: Option<String> = row.get("shortcut");
-            let parent_str: Option<String> = row.get("parent");
-            let parent_uuid = parent_str.map(ThingsId::from_trusted);
-
-            let used_ts: Option<f64> = row.get("usedDate");
-            let last_used = used_ts.and_then(|ts| {
-                let ts_i64 = safe_timestamp_convert(ts);
-                DateTime::from_timestamp(ts_i64, 0)
-            });
-
-            // Count usage by querying the TMTaskTag join table
-            let usage_count: i64 = sqlx::query_scalar(
-                "SELECT COUNT(*) FROM TMTask
-                 WHERE trashed = 0
-                 AND EXISTS (
-                     SELECT 1 FROM TMTaskTag tt
-                     WHERE tt.tasks = TMTask.uuid AND tt.tags = ?
-                 )",
-            )
-            .bind(uuid_str.as_str())
-            .fetch_one(&self.pool)
-            .await
-            .unwrap_or(0);
-
-            Ok(Some(crate::models::Tag {
-                uuid: ThingsId::from_trusted(uuid_str),
-                title,
-                shortcut,
-                parent_uuid,
-                usage_count: usage_count as u32,
-                last_used,
-            }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Find tags similar to the given title using fuzzy matching
-    ///
-    /// Returns tags sorted by similarity score (highest first)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn find_similar_tags(
-        &self,
-        title: &str,
-        min_similarity: f32,
-    ) -> ThingsResult<Vec<crate::models::TagMatch>> {
-        use crate::database::tag_utils::{calculate_similarity, get_match_type};
-
-        // Get all tags
-        let all_tags = self.get_all_tags().await?;
-
-        // Calculate similarity for each tag
-        let mut matches: Vec<crate::models::TagMatch> = all_tags
-            .into_iter()
-            .filter_map(|tag| {
-                let similarity = calculate_similarity(title, &tag.title);
-                if similarity >= min_similarity {
-                    let match_type = get_match_type(title, &tag.title, min_similarity);
-                    Some(crate::models::TagMatch {
-                        tag,
-                        similarity_score: similarity,
-                        match_type,
-                    })
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        // Sort by similarity score (highest first)
-        matches.sort_by(|a, b| {
-            b.similarity_score
-                .partial_cmp(&a.similarity_score)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        Ok(matches)
-    }
-
-    /// Search tags by partial title match
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn search_tags(&self, query: &str) -> ThingsResult<Vec<crate::models::Tag>> {
-        let rows = sqlx::query(
-            "SELECT uuid, title, shortcut, parent, usedDate
-             FROM TMTag
-             WHERE title LIKE ?
-             ORDER BY title",
-        )
-        .bind(format!("%{}%", query))
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to search tags: {e}")))?;
-
-        let mut tags = Vec::new();
-        for row in rows {
-            let uuid_str: String = row.get("uuid");
-            let title: String = row.get("title");
-            let shortcut: Option<String> = row.get("shortcut");
-            let parent_str: Option<String> = row.get("parent");
-            let parent_uuid = parent_str.map(ThingsId::from_trusted);
-
-            let used_ts: Option<f64> = row.get("usedDate");
-            let last_used = used_ts.and_then(|ts| {
-                let ts_i64 = safe_timestamp_convert(ts);
-                DateTime::from_timestamp(ts_i64, 0)
-            });
-
-            // Count usage via TMTaskTag join table
-            let usage_count: i64 = sqlx::query_scalar(
-                "SELECT COUNT(*) FROM TMTask
-                 WHERE trashed = 0
-                 AND EXISTS (
-                     SELECT 1 FROM TMTaskTag tt
-                     WHERE tt.tasks = TMTask.uuid AND tt.tags = ?
-                 )",
-            )
-            .bind(uuid_str.as_str())
-            .fetch_one(&self.pool)
-            .await
-            .unwrap_or(0);
-
-            tags.push(crate::models::Tag {
-                uuid: ThingsId::from_trusted(uuid_str),
-                title,
-                shortcut,
-                parent_uuid,
-                usage_count: usage_count as u32,
-                last_used,
-            });
-        }
-
-        Ok(tags)
-    }
-
-    /// Get all tags ordered by title
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_all_tags(&self) -> ThingsResult<Vec<crate::models::Tag>> {
-        let rows = sqlx::query(
-            "SELECT uuid, title, shortcut, parent, usedDate
-             FROM TMTag
-             ORDER BY title",
-        )
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to get all tags: {e}")))?;
-
-        let mut tags = Vec::new();
-        for row in rows {
-            let uuid_str: String = row.get("uuid");
-            let title: String = row.get("title");
-            let shortcut: Option<String> = row.get("shortcut");
-            let parent_str: Option<String> = row.get("parent");
-            let parent_uuid = parent_str.map(ThingsId::from_trusted);
-
-            let used_ts: Option<f64> = row.get("usedDate");
-            let last_used = used_ts.and_then(|ts| {
-                let ts_i64 = safe_timestamp_convert(ts);
-                DateTime::from_timestamp(ts_i64, 0)
-            });
-
-            // Count usage via TMTaskTag join table
-            let usage_count: i64 = sqlx::query_scalar(
-                "SELECT COUNT(*) FROM TMTask
-                 WHERE trashed = 0
-                 AND EXISTS (
-                     SELECT 1 FROM TMTaskTag tt
-                     WHERE tt.tasks = TMTask.uuid AND tt.tags = ?
-                 )",
-            )
-            .bind(uuid_str.as_str())
-            .fetch_one(&self.pool)
-            .await
-            .unwrap_or(0);
-
-            tags.push(crate::models::Tag {
-                uuid: ThingsId::from_trusted(uuid_str),
-                title,
-                shortcut,
-                parent_uuid,
-                usage_count: usage_count as u32,
-                last_used,
-            });
-        }
-
-        Ok(tags)
-    }
-
-    /// Get most frequently used tags
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_popular_tags(&self, limit: usize) -> ThingsResult<Vec<crate::models::Tag>> {
-        let mut all_tags = self.get_all_tags().await?;
-
-        // Sort by usage count (highest first)
-        all_tags.sort_by_key(|t| std::cmp::Reverse(t.usage_count));
-
-        // Take the top N
-        all_tags.truncate(limit);
-
-        Ok(all_tags)
-    }
-
-    /// Get recently used tags
-    ///
-    /// Recency is determined by `MAX(t.userModificationDate)` across non-trashed tasks
-    /// referencing the tag — this reflects when those tasks were last touched rather than
-    /// when they were created. Tags with no non-trashed task associations are excluded
-    /// entirely (INNER JOIN); previously the old `usedDate`-based query would have returned
-    /// them if `usedDate` were populated (it never is in practice).
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_recent_tags(&self, limit: usize) -> ThingsResult<Vec<crate::models::Tag>> {
-        // Things 3 never populates `usedDate` for tags created via its own UI or
-        // AppleScript. Instead, order by the most recent `userModificationDate` of any
-        // non-trashed task that references the tag via the TMTaskTag join table.
-        let rows = sqlx::query(
-            "SELECT tg.uuid, tg.title, tg.shortcut, tg.parent,
-                    COUNT(t.uuid) AS usage_count,
-                    MAX(t.userModificationDate) AS most_recent
-             FROM TMTag tg
-             JOIN TMTaskTag tt ON tt.tags = tg.uuid
-             JOIN TMTask t ON t.uuid = tt.tasks
-             WHERE t.trashed = 0
-             GROUP BY tg.uuid, tg.title, tg.shortcut, tg.parent
-             ORDER BY most_recent DESC
-             LIMIT ?",
-        )
-        .bind(limit as i64)
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to get recent tags: {e}")))?;
-
-        let mut tags = Vec::new();
-        for row in rows {
-            let uuid_str: String = row.get("uuid");
-            let title: String = row.get("title");
-            let shortcut: Option<String> = row.get("shortcut");
-            let parent_str: Option<String> = row.get("parent");
-            let parent_uuid = parent_str.map(ThingsId::from_trusted);
-            let usage_count: i64 = row.get("usage_count");
-
-            let most_recent_ts: Option<f64> = row.get("most_recent");
-            let last_used = most_recent_ts.and_then(|ts| {
-                let ts_i64 = safe_timestamp_convert(ts);
-                DateTime::from_timestamp(ts_i64, 0)
-            });
-
-            tags.push(crate::models::Tag {
-                uuid: ThingsId::from_trusted(uuid_str),
-                title,
-                shortcut,
-                parent_uuid,
-                usage_count: usage_count as u32,
-                last_used,
-            });
-        }
-
-        Ok(tags)
-    }
-
-    /// Create a tag with smart duplicate detection
-    ///
-    /// Returns:
-    /// - `Created`: New tag was created
-    /// - `Existing`: Exact match found (case-insensitive)
-    /// - `SimilarFound`: Similar tags found (user decision needed)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails
-    #[instrument(skip(self))]
-    pub async fn create_tag_smart(
-        &self,
-        request: crate::models::CreateTagRequest,
-    ) -> ThingsResult<crate::models::TagCreationResult> {
-        use crate::database::tag_utils::normalize_tag_title;
-        use crate::models::TagCreationResult;
-
-        // 1. Normalize the title
-        let normalized = normalize_tag_title(&request.title);
-
-        // 2. Check for exact match (case-insensitive)
-        if let Some(existing) = self.find_tag_by_normalized_title(&normalized).await? {
-            return Ok(TagCreationResult::Existing {
-                tag: existing,
-                is_new: false,
-            });
-        }
-
-        // 3. Find similar tags (fuzzy matching with 80% threshold)
-        let similar_tags = self.find_similar_tags(&normalized, 0.8).await?;
-
-        // 4. If similar tags found, return them for user decision
-        if !similar_tags.is_empty() {
-            return Ok(TagCreationResult::SimilarFound {
-                similar_tags,
-                requested_title: request.title,
-            });
-        }
-
-        // 5. No duplicates, safe to create
-        let id = ThingsId::new_things_native();
-
-        sqlx::query(
-            "INSERT INTO TMTag (uuid, title, shortcut, parent, usedDate, `index`) \
-             VALUES (?, ?, ?, ?, NULL, 0)",
-        )
-        .bind(id.as_str())
-        .bind(&request.title)
-        .bind(request.shortcut.as_ref())
-        .bind(request.parent_uuid.map(|u| u.into_string()))
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to create tag: {e}")))?;
-
-        info!("Created tag with UUID: {}", id);
-        Ok(TagCreationResult::Created {
-            uuid: id,
-            is_new: true,
-        })
-    }
-
-    /// Create tag forcefully (skip duplicate check)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails
-    #[instrument(skip(self))]
-    pub async fn create_tag_force(
-        &self,
-        request: crate::models::CreateTagRequest,
-    ) -> ThingsResult<ThingsId> {
-        let id = ThingsId::new_things_native();
-
-        sqlx::query(
-            "INSERT INTO TMTag (uuid, title, shortcut, parent, usedDate, `index`) \
-             VALUES (?, ?, ?, ?, NULL, 0)",
-        )
-        .bind(id.as_str())
-        .bind(&request.title)
-        .bind(request.shortcut.as_ref())
-        .bind(request.parent_uuid.map(|u| u.into_string()))
-        .execute(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to create tag: {e}")))?;
-
-        info!("Forcefully created tag with UUID: {}", id);
-        Ok(id)
-    }
-
-    /// Update a tag
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the tag doesn't exist or database operation fails
-    #[instrument(skip(self))]
-    pub async fn update_tag(&self, request: crate::models::UpdateTagRequest) -> ThingsResult<()> {
-        use crate::database::tag_utils::normalize_tag_title;
-
-        // Verify tag exists
-        let existing = self
-            .find_tag_by_normalized_title(request.uuid.as_str())
-            .await?;
-        if existing.is_none() {
-            // Try by UUID
-            let row = sqlx::query("SELECT 1 FROM TMTag WHERE uuid = ?")
-                .bind(request.uuid.as_str())
-                .fetch_optional(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to validate tag: {e}")))?;
-
-            if row.is_none() {
-                return Err(ThingsError::unknown(format!(
-                    "Tag not found: {}",
-                    request.uuid
-                )));
-            }
-        }
-
-        // If renaming, check for duplicates with new name
-        if let Some(new_title) = &request.title {
-            let normalized = normalize_tag_title(new_title);
-            if let Some(duplicate) = self.find_tag_by_normalized_title(&normalized).await? {
-                if duplicate.uuid != request.uuid {
-                    return Err(ThingsError::unknown(format!(
-                        "Tag with title '{}' already exists",
-                        new_title
-                    )));
-                }
-            }
-        }
-
-        // Build dynamic UPDATE query
-        let mut updates = Vec::new();
-        let mut params: Vec<String> = Vec::new();
-
-        if let Some(title) = &request.title {
-            updates.push("title = ?");
-            params.push(title.clone());
-        }
-        if let Some(shortcut) = &request.shortcut {
-            updates.push("shortcut = ?");
-            params.push(shortcut.clone());
-        }
-        if let Some(parent_uuid) = request.parent_uuid {
-            updates.push("parent = ?");
-            params.push(parent_uuid.into_string());
-        }
-
-        if updates.is_empty() {
-            return Ok(()); // Nothing to update
-        }
-
-        let sql = format!("UPDATE TMTag SET {} WHERE uuid = ?", updates.join(", "));
-        params.push(request.uuid.as_str().to_string());
-
-        let mut query = sqlx::query(&sql);
-        for param in params {
-            query = query.bind(param);
-        }
-
-        query
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to update tag: {e}")))?;
-
-        info!("Updated tag with UUID: {}", request.uuid);
-        Ok(())
-    }
-
-    /// Delete a tag
-    ///
-    /// # Arguments
-    ///
-    /// * `uuid` - UUID of the tag to delete
-    /// * `remove_from_tasks` - If true, removes tag from all tasks' cachedTags
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails
-    #[instrument(skip(self))]
-    pub async fn delete_tag(&self, id: &ThingsId, remove_from_tasks: bool) -> ThingsResult<()> {
-        // Get the tag title before deletion
-        let tag = self.find_tag_by_normalized_title(id.as_str()).await?;
-
-        if tag.is_none() {
-            // Try by UUID directly
-            let row = sqlx::query("SELECT title FROM TMTag WHERE uuid = ?")
-                .bind(id.as_str())
-                .fetch_optional(&self.pool)
-                .await
-                .map_err(|e| ThingsError::unknown(format!("Failed to find tag: {e}")))?;
-
-            if row.is_none() {
-                return Err(ThingsError::unknown(format!("Tag not found: {}", id)));
-            }
-        }
-
-        if remove_from_tasks {
-            // TODO: Implement updating all tasks' cachedTags to remove this tag
-            // This requires parsing and re-serializing the JSON arrays
-            info!("Removing tag {} from all tasks (not yet implemented)", id);
-        }
-
-        // Delete the tag
-        sqlx::query("DELETE FROM TMTag WHERE uuid = ?")
-            .bind(id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to delete tag: {e}")))?;
-
-        info!("Deleted tag with UUID: {}", id);
-        Ok(())
-    }
-
-    /// Merge two tags (combine source into target)
-    ///
-    /// # Arguments
-    ///
-    /// * `source_uuid` - UUID of tag to merge from (will be deleted)
-    /// * `target_uuid` - UUID of tag to merge into (will remain)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if either tag doesn't exist or database operation fails
-    #[instrument(skip(self))]
-    pub async fn merge_tags(&self, source_id: &ThingsId, target_id: &ThingsId) -> ThingsResult<()> {
-        // Verify both tags exist
-        let source_row = sqlx::query("SELECT title FROM TMTag WHERE uuid = ?")
-            .bind(source_id.as_str())
-            .fetch_optional(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to find source tag: {e}")))?;
-
-        if source_row.is_none() {
-            return Err(ThingsError::unknown(format!(
-                "Source tag not found: {}",
-                source_id
-            )));
-        }
-
-        let target_row = sqlx::query("SELECT title FROM TMTag WHERE uuid = ?")
-            .bind(target_id.as_str())
-            .fetch_optional(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to find target tag: {e}")))?;
-
-        if target_row.is_none() {
-            return Err(ThingsError::unknown(format!(
-                "Target tag not found: {}",
-                target_id
-            )));
-        }
-
-        // TODO: Implement updating all tasks' cachedTags to replace source tag with target tag
-        // This requires parsing and re-serializing the JSON arrays
-        info!(
-            "Merging tag {} into {} (tag replacement in tasks not yet fully implemented)",
-            source_id, target_id
-        );
-
-        // Update usedDate on target if source was used more recently
-        let now = Utc::now().timestamp() as f64;
-        sqlx::query("UPDATE TMTag SET usedDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(target_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to update target tag: {e}")))?;
-
-        // Delete source tag
-        sqlx::query("DELETE FROM TMTag WHERE uuid = ?")
-            .bind(source_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to delete source tag: {e}")))?;
-
-        info!("Merged tag {} into {}", source_id, target_id);
-        Ok(())
-    }
 
     // ========================================================================
     // TAG ASSIGNMENT OPERATIONS
     // ========================================================================
 
-    /// Add a tag to a task (with duplicate prevention)
-    ///
-    /// Returns:
-    /// - `Assigned`: Tag was successfully assigned
-    /// - `Suggestions`: Similar tags found (user decision needed)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task doesn't exist or database operation fails
-    #[instrument(skip(self))]
-    pub async fn add_tag_to_task(
-        &self,
-        task_id: &ThingsId,
-        tag_title: &str,
-    ) -> ThingsResult<crate::models::TagAssignmentResult> {
-        use crate::database::tag_utils::normalize_tag_title;
-        use crate::models::TagAssignmentResult;
-
-        // 1. Verify task exists
-        validators::validate_task_exists(&self.pool, task_id).await?;
-
-        // 2. Normalize and find tag
-        let normalized = normalize_tag_title(tag_title);
-
-        // 3. Check for exact match first
-        let tag = if let Some(existing_tag) = self.find_tag_by_normalized_title(&normalized).await?
-        {
-            existing_tag
-        } else {
-            // 4. Find similar tags
-            let similar_tags = self.find_similar_tags(&normalized, 0.8).await?;
-
-            if !similar_tags.is_empty() {
-                return Ok(TagAssignmentResult::Suggestions { similar_tags });
-            }
-
-            // 5. No existing tag found, create new one
-            let request = crate::models::CreateTagRequest {
-                title: tag_title.to_string(),
-                shortcut: None,
-                parent_uuid: None,
-            };
-            let _uuid = self.create_tag_force(request).await?;
-
-            // Fetch the newly created tag
-            self.find_tag_by_normalized_title(&normalized)
-                .await?
-                .ok_or_else(|| ThingsError::unknown("Failed to retrieve newly created tag"))?
-        };
-
-        // 6. Insert into TMTaskTag (idempotent)
-        let now = Utc::now().timestamp() as f64;
-        sqlx::query("INSERT OR IGNORE INTO TMTaskTag (tasks, tags) VALUES (?, ?)")
-            .bind(task_id.as_str())
-            .bind(tag.uuid.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to assign tag: {e}")))?;
-
-        // 7. Update task modification date
-        sqlx::query("UPDATE TMTask SET userModificationDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(task_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| {
-                ThingsError::unknown(format!("Failed to update modification date: {e}"))
-            })?;
-
-        // 8. Update tag's usedDate
-        sqlx::query("UPDATE TMTag SET usedDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(tag.uuid.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to update tag usedDate: {e}")))?;
-
-        info!("Added tag '{}' to task {}", tag.title, task_id);
-
-        Ok(TagAssignmentResult::Assigned { tag_uuid: tag.uuid })
-    }
-
-    /// Remove a tag from a task
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task doesn't exist or database operation fails
-    #[instrument(skip(self))]
-    pub async fn remove_tag_from_task(
-        &self,
-        task_id: &ThingsId,
-        tag_title: &str,
-    ) -> ThingsResult<()> {
-        use crate::database::tag_utils::normalize_tag_title;
-
-        // 1. Verify task exists
-        validators::validate_task_exists(&self.pool, task_id).await?;
-
-        // 2. Find the tag UUID
-        let normalized = normalize_tag_title(tag_title);
-        let Some(tag) = self.find_tag_by_normalized_title(&normalized).await? else {
-            return Ok(()); // Tag doesn't exist, nothing to remove
-        };
-
-        // 3. Delete from TMTaskTag
-        let now = Utc::now().timestamp() as f64;
-        let result = sqlx::query("DELETE FROM TMTaskTag WHERE tasks = ? AND tags = ?")
-            .bind(task_id.as_str())
-            .bind(tag.uuid.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to remove tag assignment: {e}")))?;
-
-        if result.rows_affected() > 0 {
-            // 4. Update task modification date
-            sqlx::query("UPDATE TMTask SET userModificationDate = ? WHERE uuid = ?")
-                .bind(now)
-                .bind(task_id.as_str())
-                .execute(&self.pool)
-                .await
-                .map_err(|e| {
-                    ThingsError::unknown(format!("Failed to update modification date: {e}"))
-                })?;
-
-            info!("Removed tag '{}' from task {}", tag_title, task_id);
-        }
-
-        Ok(())
-    }
-
-    /// Replace all tags on a task (with duplicate prevention)
-    ///
-    /// Returns any tag titles that had similar matches for user confirmation
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the task doesn't exist or database operation fails
-    #[instrument(skip(self))]
-    pub async fn set_task_tags(
-        &self,
-        task_id: &ThingsId,
-        tag_titles: Vec<String>,
-    ) -> ThingsResult<Vec<crate::models::TagMatch>> {
-        use crate::database::tag_utils::normalize_tag_title;
-
-        // 1. Verify task exists
-        validators::validate_task_exists(&self.pool, task_id).await?;
-
-        let mut resolved_tags = Vec::new();
-        let mut suggestions = Vec::new();
-
-        // 2. Resolve each tag title
-        for title in tag_titles {
-            let normalized = normalize_tag_title(&title);
-
-            // Try to find exact match
-            if let Some(existing_tag) = self.find_tag_by_normalized_title(&normalized).await? {
-                resolved_tags.push(existing_tag.title);
-            } else {
-                // Check for similar tags
-                let similar_tags = self.find_similar_tags(&normalized, 0.8).await?;
-
-                if !similar_tags.is_empty() {
-                    suggestions.extend(similar_tags);
-                }
-
-                // Use the requested title anyway (will create if needed)
-                resolved_tags.push(title);
-            }
-        }
-
-        // 3. For any tags that don't exist yet, create them
-        for title in &resolved_tags {
-            let normalized = normalize_tag_title(title);
-            if self
-                .find_tag_by_normalized_title(&normalized)
-                .await?
-                .is_none()
-            {
-                let request = crate::models::CreateTagRequest {
-                    title: title.clone(),
-                    shortcut: None,
-                    parent_uuid: None,
-                };
-                self.create_tag_force(request).await?;
-            }
-        }
-
-        // 4. Delete existing tag assignments for this task
-        let now = Utc::now().timestamp() as f64;
-        sqlx::query("DELETE FROM TMTaskTag WHERE tasks = ?")
-            .bind(task_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to clear task tags: {e}")))?;
-
-        // 5. Insert new tag assignments
-        for title in &resolved_tags {
-            let normalized = normalize_tag_title(title);
-            if let Some(tag) = self.find_tag_by_normalized_title(&normalized).await? {
-                sqlx::query("INSERT OR IGNORE INTO TMTaskTag (tasks, tags) VALUES (?, ?)")
-                    .bind(task_id.as_str())
-                    .bind(tag.uuid.as_str())
-                    .execute(&self.pool)
-                    .await
-                    .map_err(|e| ThingsError::unknown(format!("Failed to assign tag: {e}")))?;
-            }
-        }
-
-        // Update task modification date
-        sqlx::query("UPDATE TMTask SET userModificationDate = ? WHERE uuid = ?")
-            .bind(now)
-            .bind(task_id.as_str())
-            .execute(&self.pool)
-            .await
-            .map_err(|e| {
-                ThingsError::unknown(format!("Failed to update modification date: {e}"))
-            })?;
-
-        // 6. Update usedDate for all tags
-        for title in &resolved_tags {
-            let normalized = normalize_tag_title(title);
-            if let Some(tag) = self.find_tag_by_normalized_title(&normalized).await? {
-                sqlx::query("UPDATE TMTag SET usedDate = ? WHERE uuid = ?")
-                    .bind(now)
-                    .bind(tag.uuid.as_str())
-                    .execute(&self.pool)
-                    .await
-                    .map_err(|e| {
-                        ThingsError::unknown(format!("Failed to update tag usedDate: {e}"))
-                    })?;
-            }
-        }
-
-        info!("Set tags on task {} to: {:?}", task_id, resolved_tags);
-        Ok(suggestions)
-    }
-
     // ========================================================================
     // TAG AUTO-COMPLETION & ANALYTICS
     // ========================================================================
 
-    /// Get tag completions for partial input
-    ///
-    /// Returns tags sorted by:
-    /// 1. Exact prefix matches (prioritized)
-    /// 2. Contains matches
-    /// 3. Fuzzy matches
-    /// Within each category, sorted by usage frequency
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn get_tag_completions(
-        &self,
-        partial_input: &str,
-        limit: usize,
-    ) -> ThingsResult<Vec<crate::models::TagCompletion>> {
-        use crate::database::tag_utils::{calculate_similarity, normalize_tag_title};
-
-        let normalized_input = normalize_tag_title(partial_input);
-        let all_tags = self.get_all_tags().await?;
-
-        let mut completions: Vec<crate::models::TagCompletion> = all_tags
-            .into_iter()
-            .filter_map(|tag| {
-                let normalized_tag = normalize_tag_title(&tag.title);
-
-                // Calculate score based on match type
-                let score = if normalized_tag.starts_with(&normalized_input) {
-                    // Exact prefix match: highest priority
-                    3.0 + (tag.usage_count as f32 / 100.0)
-                } else if normalized_tag.contains(&normalized_input) {
-                    // Contains match: medium priority
-                    2.0 + (tag.usage_count as f32 / 100.0)
-                } else {
-                    // Fuzzy match: lower priority
-                    let similarity = calculate_similarity(partial_input, &tag.title);
-                    if similarity >= 0.6 {
-                        similarity + (tag.usage_count as f32 / 1000.0)
-                    } else {
-                        return None; // Not similar enough
-                    }
-                };
-
-                Some(crate::models::TagCompletion { tag, score })
-            })
-            .collect();
-
-        // Sort by score (highest first)
-        completions.sort_by(|a, b| {
-            b.score
-                .partial_cmp(&a.score)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        // Take the top N
-        completions.truncate(limit);
-
-        Ok(completions)
-    }
-
-    /// Get detailed statistics for a tag
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the tag doesn't exist or database query fails
-    #[instrument(skip(self))]
-    pub async fn get_tag_statistics(
-        &self,
-        id: &ThingsId,
-    ) -> ThingsResult<crate::models::TagStatistics> {
-        // Get the tag
-        let tag_row = sqlx::query("SELECT title FROM TMTag WHERE uuid = ?")
-            .bind(id.as_str())
-            .fetch_optional(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to find tag: {e}")))?;
-
-        let title: String = tag_row
-            .ok_or_else(|| ThingsError::unknown(format!("Tag not found: {}", id)))?
-            .get("title");
-
-        // Get all tasks using this tag via TMTaskTag join table
-        let task_rows = sqlx::query(
-            "SELECT tt.tasks AS uuid FROM TMTaskTag tt
-             JOIN TMTask t ON t.uuid = tt.tasks
-             WHERE tt.tags = ? AND t.trashed = 0",
-        )
-        .bind(id.as_str())
-        .fetch_all(&self.pool)
-        .await
-        .map_err(|e| ThingsError::unknown(format!("Failed to query tasks with tag: {e}")))?;
-
-        let task_uuids: Vec<ThingsId> = task_rows
-            .iter()
-            .map(|row| ThingsId::from_trusted(row.get::<String, _>("uuid")))
-            .collect();
-
-        let usage_count = task_uuids.len() as u32;
-
-        // Find related tags (tags that frequently appear with this tag)
-        let mut related_tags: std::collections::HashMap<String, u32> =
-            std::collections::HashMap::new();
-
-        for task_uuid in &task_uuids {
-            let rows = sqlx::query(
-                "SELECT DISTINCT tg.title FROM TMTaskTag tt
-                 JOIN TMTag tg ON tg.uuid = tt.tags
-                 WHERE tt.tasks = ?",
-            )
-            .bind(task_uuid.as_str())
-            .fetch_all(&self.pool)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to fetch task tags: {e}")))?;
-
-            for row in rows {
-                let tag_title: String = row.get("title");
-                if tag_title != title {
-                    *related_tags.entry(tag_title).or_insert(0) += 1;
-                }
-            }
-        }
-
-        // Sort related tags by co-occurrence count
-        let mut related_vec: Vec<(String, u32)> = related_tags.into_iter().collect();
-        related_vec.sort_by_key(|r| std::cmp::Reverse(r.1));
-
-        Ok(crate::models::TagStatistics {
-            uuid: id.clone(),
-            title,
-            usage_count,
-            task_uuids,
-            related_tags: related_vec,
-        })
-    }
-
-    /// Find duplicate or highly similar tags
-    ///
-    /// Returns pairs of tags that are similar above the threshold
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails
-    #[instrument(skip(self))]
-    pub async fn find_duplicate_tags(
-        &self,
-        min_similarity: f32,
-    ) -> ThingsResult<Vec<crate::models::TagPair>> {
-        use crate::database::tag_utils::calculate_similarity;
-
-        let all_tags = self.get_all_tags().await?;
-        let mut pairs = Vec::new();
-
-        // Compare each tag with every other tag
-        for i in 0..all_tags.len() {
-            for j in (i + 1)..all_tags.len() {
-                let tag1 = &all_tags[i];
-                let tag2 = &all_tags[j];
-
-                let similarity = calculate_similarity(&tag1.title, &tag2.title);
-
-                if similarity >= min_similarity {
-                    pairs.push(crate::models::TagPair {
-                        tag1: tag1.clone(),
-                        tag2: tag2.clone(),
-                        similarity,
-                    });
-                }
-            }
-        }
-
-        // Sort by similarity (highest first)
-        pairs.sort_by(|a, b| {
-            b.similarity
-                .partial_cmp(&a.similarity)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        Ok(pairs)
-    }
-
     // ============================================================================
     // Bulk Operations
     // ============================================================================
-
-    /// Maximum number of tasks that can be processed in a single bulk operation
-    /// This prevents abuse and ensures reasonable transaction sizes
-    const MAX_BULK_BATCH_SIZE: usize = 1000;
-
-    /// Move multiple tasks to a project or area (transactional)
-    ///
-    /// All tasks must exist and be valid, or the entire operation will be rolled back.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - Task UUIDs array is empty
-    /// - Neither project_uuid nor area_uuid is specified
-    /// - Target project or area doesn't exist
-    /// - Any task UUID is invalid or doesn't exist
-    /// - Database operation fails
-    #[instrument(skip(self))]
-    pub async fn bulk_move(
-        &self,
-        request: crate::models::BulkMoveRequest,
-    ) -> ThingsResult<crate::models::BulkOperationResult> {
-        // Validation
-        if request.task_uuids.is_empty() {
-            return Err(ThingsError::validation("Task UUIDs cannot be empty"));
-        }
-        if request.task_uuids.len() > Self::MAX_BULK_BATCH_SIZE {
-            return Err(ThingsError::validation(format!(
-                "Batch size {} exceeds maximum of {}",
-                request.task_uuids.len(),
-                Self::MAX_BULK_BATCH_SIZE
-            )));
-        }
-        if request.project_uuid.is_none() && request.area_uuid.is_none() {
-            return Err(ThingsError::validation(
-                "Must specify either project_uuid or area_uuid",
-            ));
-        }
-
-        // Validate target project/area exists
-        if let Some(project_uuid) = &request.project_uuid {
-            validators::validate_project_exists(&self.pool, project_uuid).await?;
-        }
-        if let Some(area_uuid) = &request.area_uuid {
-            validators::validate_area_exists(&self.pool, area_uuid).await?;
-        }
-
-        // Begin transaction
-        let mut tx = self
-            .pool
-            .begin()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to begin transaction: {e}")))?;
-
-        // Validate all tasks exist in a single batch query (prevent N+1)
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "SELECT uuid FROM TMTask WHERE uuid IN ({}) AND trashed = 0",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str);
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        let found_uuids: Vec<String> = query
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to validate tasks: {e}")))?
-            .iter()
-            .map(|row| row.get("uuid"))
-            .collect();
-
-        // Check if any UUIDs were not found
-        if found_uuids.len() != request.task_uuids.len() {
-            // Find the first missing UUID for error reporting
-            for id in &request.task_uuids {
-                if !found_uuids.contains(&id.to_string()) {
-                    tx.rollback().await.ok();
-                    return Err(ThingsError::TaskNotFound {
-                        uuid: id.to_string(),
-                    });
-                }
-            }
-        }
-
-        // Batch update
-        let now = Utc::now().timestamp() as f64;
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "UPDATE TMTask SET project = ?, area = ?, userModificationDate = ? WHERE uuid IN ({})",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str)
-            .bind(request.project_uuid.map(|u| u.into_string()))
-            .bind(request.area_uuid.map(|u| u.into_string()))
-            .bind(now);
-
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        query
-            .execute(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to bulk move tasks: {e}")))?;
-
-        // Commit transaction
-        tx.commit()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to commit transaction: {e}")))?;
-
-        info!("Bulk moved {} task(s)", request.task_uuids.len());
-        Ok(crate::models::BulkOperationResult {
-            success: true,
-            processed_count: request.task_uuids.len(),
-            message: format!("Successfully moved {} task(s)", request.task_uuids.len()),
-        })
-    }
-
-    /// Update dates for multiple tasks with validation (transactional)
-    ///
-    /// All tasks must exist and dates must be valid, or the entire operation will be rolled back.
-    /// Validates that deadline >= start_date for each task after merging with existing dates.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - Task UUIDs array is empty
-    /// - Any task UUID is invalid or doesn't exist
-    /// - Date range validation fails (deadline before start_date)
-    /// - Database operation fails
-    #[instrument(skip(self))]
-    pub async fn bulk_update_dates(
-        &self,
-        request: crate::models::BulkUpdateDatesRequest,
-    ) -> ThingsResult<crate::models::BulkOperationResult> {
-        use crate::database::{safe_things_date_to_naive_date, validate_date_range};
-
-        // Validation
-        if request.task_uuids.is_empty() {
-            return Err(ThingsError::validation("Task UUIDs cannot be empty"));
-        }
-        if request.task_uuids.len() > Self::MAX_BULK_BATCH_SIZE {
-            return Err(ThingsError::validation(format!(
-                "Batch size {} exceeds maximum of {}",
-                request.task_uuids.len(),
-                Self::MAX_BULK_BATCH_SIZE
-            )));
-        }
-
-        // Validate date range if both are provided
-        if let (Some(start), Some(deadline)) = (request.start_date, request.deadline) {
-            validate_date_range(Some(start), Some(deadline))?;
-        }
-
-        // Begin transaction
-        let mut tx = self
-            .pool
-            .begin()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to begin transaction: {e}")))?;
-
-        // Validate all tasks exist and check merged date validity in a single batch query
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "SELECT uuid, startDate, deadline FROM TMTask WHERE uuid IN ({}) AND trashed = 0",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str);
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        let rows = query
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to validate tasks: {e}")))?;
-
-        // Check if all UUIDs were found
-        if rows.len() != request.task_uuids.len() {
-            // Find the first missing UUID for error reporting
-            let found_uuids: Vec<String> = rows.iter().map(|row| row.get("uuid")).collect();
-            for id in &request.task_uuids {
-                if !found_uuids.contains(&id.to_string()) {
-                    tx.rollback().await.ok();
-                    return Err(ThingsError::TaskNotFound {
-                        uuid: id.to_string(),
-                    });
-                }
-            }
-        }
-
-        // Validate merged dates for all tasks
-        for row in &rows {
-            let current_start: Option<i64> = row.get("startDate");
-            let current_deadline: Option<i64> = row.get("deadline");
-
-            let final_start = if request.clear_start_date {
-                None
-            } else if let Some(new_start) = request.start_date {
-                Some(new_start)
-            } else {
-                current_start.and_then(|ts| safe_things_date_to_naive_date(ts).ok())
-            };
-
-            let final_deadline = if request.clear_deadline {
-                None
-            } else if let Some(new_deadline) = request.deadline {
-                Some(new_deadline)
-            } else {
-                current_deadline.and_then(|ts| safe_things_date_to_naive_date(ts).ok())
-            };
-
-            validate_date_range(final_start, final_deadline)?;
-        }
-
-        // Build and execute batch update
-        let now = Utc::now().timestamp() as f64;
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-
-        let start_date_value = if request.clear_start_date {
-            None
-        } else {
-            request.start_date.map(naive_date_to_things_timestamp)
-        };
-
-        let deadline_value = if request.clear_deadline {
-            None
-        } else {
-            request.deadline.map(naive_date_to_things_timestamp)
-        };
-
-        let query_str = format!(
-            "UPDATE TMTask SET startDate = ?, deadline = ?, userModificationDate = ? WHERE uuid IN ({})",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str)
-            .bind(start_date_value)
-            .bind(deadline_value)
-            .bind(now);
-
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        query
-            .execute(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to bulk update dates: {e}")))?;
-
-        tx.commit()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to commit transaction: {e}")))?;
-
-        info!(
-            "Bulk updated dates for {} task(s)",
-            request.task_uuids.len()
-        );
-        Ok(crate::models::BulkOperationResult {
-            success: true,
-            processed_count: request.task_uuids.len(),
-            message: format!(
-                "Successfully updated dates for {} task(s)",
-                request.task_uuids.len()
-            ),
-        })
-    }
-
-    /// Complete multiple tasks (transactional)
-    ///
-    /// All tasks must exist, or the entire operation will be rolled back.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - Task UUIDs array is empty
-    /// - Any task UUID is invalid or doesn't exist
-    /// - Database operation fails
-    #[instrument(skip(self))]
-    pub async fn bulk_complete(
-        &self,
-        request: crate::models::BulkCompleteRequest,
-    ) -> ThingsResult<crate::models::BulkOperationResult> {
-        // Validation
-        if request.task_uuids.is_empty() {
-            return Err(ThingsError::validation("Task UUIDs cannot be empty"));
-        }
-        if request.task_uuids.len() > Self::MAX_BULK_BATCH_SIZE {
-            return Err(ThingsError::validation(format!(
-                "Batch size {} exceeds maximum of {}",
-                request.task_uuids.len(),
-                Self::MAX_BULK_BATCH_SIZE
-            )));
-        }
-
-        // Begin transaction
-        let mut tx = self
-            .pool
-            .begin()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to begin transaction: {e}")))?;
-
-        // Validate all tasks exist in a single batch query (prevent N+1)
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "SELECT uuid FROM TMTask WHERE uuid IN ({}) AND trashed = 0",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str);
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        let found_uuids: Vec<String> = query
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to validate tasks: {e}")))?
-            .iter()
-            .map(|row| row.get("uuid"))
-            .collect();
-
-        // Check if any UUIDs were not found
-        if found_uuids.len() != request.task_uuids.len() {
-            // Find the first missing UUID for error reporting
-            for id in &request.task_uuids {
-                if !found_uuids.contains(&id.to_string()) {
-                    tx.rollback().await.ok();
-                    return Err(ThingsError::TaskNotFound {
-                        uuid: id.to_string(),
-                    });
-                }
-            }
-        }
-
-        // Batch update: mark as completed
-        let now = Utc::now().timestamp() as f64;
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "UPDATE TMTask SET status = 3, stopDate = ?, userModificationDate = ? WHERE uuid IN ({})",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str).bind(now).bind(now);
-
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        query
-            .execute(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to bulk complete tasks: {e}")))?;
-
-        // Commit transaction
-        tx.commit()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to commit transaction: {e}")))?;
-
-        info!("Bulk completed {} task(s)", request.task_uuids.len());
-        Ok(crate::models::BulkOperationResult {
-            success: true,
-            processed_count: request.task_uuids.len(),
-            message: format!(
-                "Successfully completed {} task(s)",
-                request.task_uuids.len()
-            ),
-        })
-    }
-
-    /// Delete multiple tasks (soft delete, transactional)
-    ///
-    /// All tasks must exist, or the entire operation will be rolled back.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - Task UUIDs array is empty
-    /// - Any task UUID is invalid or doesn't exist
-    /// - Database operation fails
-    #[instrument(skip(self))]
-    pub async fn bulk_delete(
-        &self,
-        request: crate::models::BulkDeleteRequest,
-    ) -> ThingsResult<crate::models::BulkOperationResult> {
-        // Validation
-        if request.task_uuids.is_empty() {
-            return Err(ThingsError::validation("Task UUIDs cannot be empty"));
-        }
-        if request.task_uuids.len() > Self::MAX_BULK_BATCH_SIZE {
-            return Err(ThingsError::validation(format!(
-                "Batch size {} exceeds maximum of {}",
-                request.task_uuids.len(),
-                Self::MAX_BULK_BATCH_SIZE
-            )));
-        }
-
-        // Begin transaction
-        let mut tx = self
-            .pool
-            .begin()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to begin transaction: {e}")))?;
-
-        // Validate all tasks exist in a single batch query (prevent N+1)
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "SELECT uuid FROM TMTask WHERE uuid IN ({}) AND trashed = 0",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str);
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        let found_uuids: Vec<String> = query
-            .fetch_all(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to validate tasks: {e}")))?
-            .iter()
-            .map(|row| row.get("uuid"))
-            .collect();
-
-        // Check if any UUIDs were not found
-        if found_uuids.len() != request.task_uuids.len() {
-            // Find the first missing UUID for error reporting
-            for id in &request.task_uuids {
-                if !found_uuids.contains(&id.to_string()) {
-                    tx.rollback().await.ok();
-                    return Err(ThingsError::TaskNotFound {
-                        uuid: id.to_string(),
-                    });
-                }
-            }
-        }
-
-        // Batch update: soft delete
-        let now = Utc::now().timestamp() as f64;
-        let placeholders = request
-            .task_uuids
-            .iter()
-            .map(|_| "?")
-            .collect::<Vec<_>>()
-            .join(",");
-        let query_str = format!(
-            "UPDATE TMTask SET trashed = 1, userModificationDate = ? WHERE uuid IN ({})",
-            placeholders
-        );
-
-        let mut query = sqlx::query(&query_str).bind(now);
-
-        for id in &request.task_uuids {
-            query = query.bind(id.as_str());
-        }
-
-        query
-            .execute(&mut *tx)
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to bulk delete tasks: {e}")))?;
-
-        // Commit transaction
-        tx.commit()
-            .await
-            .map_err(|e| ThingsError::unknown(format!("Failed to commit transaction: {e}")))?;
-
-        info!("Bulk deleted {} task(s)", request.task_uuids.len());
-        Ok(crate::models::BulkOperationResult {
-            success: true,
-            processed_count: request.task_uuids.len(),
-            message: format!("Successfully deleted {} task(s)", request.task_uuids.len()),
-        })
-    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::models::{
+        Area, BulkCompleteRequest, BulkDeleteRequest, BulkMoveRequest, BulkUpdateDatesRequest,
+        CreateProjectRequest, CreateTaskRequest, DeleteChildHandling, TaskStatus, TaskType,
+        ThingsId, UpdateTaskRequest,
+    };
     use std::time::Duration;
     use tempfile::{NamedTempFile, TempDir};
