+++ libs/things3-core/src/database/mod.rs
@@ -4,8 +4,10 @@ pub mod conversions;
 mod core;
 pub mod date_utils;
 pub mod mappers;
+mod mutations;
 pub mod path_discovery;
 pub mod pool;
+mod queries;
 pub mod query_builders;
 pub mod stats;
 pub mod tag_utils;
