+++ libs/things3-core/src/database/mutations/mod.rs
@@ -1,0 +1,5 @@
+mod areas;
+mod bulk;
+mod projects;
+mod tags;
+mod tasks;
