+++ libs/things3-core/src/database/queries/mod.rs
@@ -1,0 +1,4 @@
+mod areas;
+mod projects;
+mod tags;
+mod tasks;
